actualizacion

+++ app/Http/Controllers/Mod01_ProduccionController.php
@@ -8,6 +8,7 @@ use App\Modelos\MOD01\LOGOF;
 use App\Modelos\MOD01\LOGOT;
 use App\OP;
 use App\SAP;
+use App\SAPi;
 use App\User;
 use Auth;
 use DB;
@@ -223,9 +224,7 @@ class Mod01_ProduccionController extends Controller
                     $est_Av = $t_user->U_CP_CT;
                     $Fil_Est = explode(",", $est_Av); //ARRAY SIMPLE
                     $rutasConNombres = self::getNombresRutas($Fil_Est); //ARRAY LLAVE VALOR
-                    // array_pop($rutasConNombres);
-                    //$ruts1=DB::select("SELECT Name From [@PL_RUTAS] where Code=".Input::get('OP_us'));
-
+                     
                     return view('Mod01_Produccion.traslados', ['rutasConNombres' => $rutasConNombres, 't_user' => $t_user, 'est_Av' => $est_Av, 'Fil_Est' => $Fil_Est, 'actividades' => $actividades, 'ultimo' => count($actividades), 't_user' => $t_user]);
                 } else {
                     return redirect()->back()->withErrors(array('message' => 'Error de autenticación.'));
@@ -289,11 +288,11 @@ class Mod01_ProduccionController extends Controller
                     return redirect()->route('home');
                     //dd('getOPss');
                 }
-                $Codes = OP::where('U_DocEntry', $op)->get();
+                $Codes = OP::where('U_DocEntry', $op)->get();              
             } else{
                 $Codes = [];
             }
-        
+       
             Session::flash('usertraslados', 2); //evita que salga el modal
 
             if (count($Codes) > 0) {
@@ -443,7 +442,7 @@ class Mod01_ProduccionController extends Controller
                  't_user' => $t_user, 
                  'ofs' => $one, 
                  'op' => $op, 
-                 'pedido' => $pedido, 
+                 'pedido' => $pedido,  
                  'HisOrden' => $HisOrden]);
             } else if (count($Codes) == 0 && Session::has('recibo')){ // se hizo un recibo
                 return view('Mod01_Produccion.traslados',
@@ -1187,12 +1186,17 @@ public function terminarOP(){
 
    $rates = DB::table('ORTT')->where('RateDate', date('d-m-Y'))->get();
         if (count($rates) >= 3) {
-            $result = \App\SAPi::ReciboProduccion((int)Input::get('orden'), (int)Input::get('cant'));
-            if (strpos($result, 'Recibo') !== false) {
+            $Code_actual = OP::find(Input::get('code'));
+            $orden_owor = DB::table('OWOR')->where('DocNum', Input::get('orden'))->first();
+            if (($orden_owor->PlannedQty) >= (floatval ( $orden_owor->CmpltQty) + floatval ( Input::get('cant')))) {
+                $apellido = Self::getApellidoPaternoUsuario(explode(' ',Auth::user()->lastName));
+                $usuario_reporta = explode(' ', Auth::user()->firstName)[0].' '.$apellido;
+                $result = SAPi::ReciboProduccion(Input::get('orden'), $orden_owor->Warehouse, Input::get('cant'), "Reportado por: ".$usuario_reporta, "Recibo de producción");
+                if (strpos($result, 'Recibo') !== false) {
                 //agregar linea en LOGOF del avance
-                    $dt = new \DateTime();
+                    $dt = date('Ymd h:m:s');
                     $user = User::find(Input::get('userId'));
-                    $Code_actual = OP::find(Input::get('code'));
+//$Code_actual = OP::find(Input::get('code'));
 
                     $Con_Logof = DB::select('select max (CONVERT(INT,Code)) as Code FROM  [@CP_LOGOF]');
                     $log = new LOGOF();
@@ -1208,7 +1212,12 @@ public function terminarOP(){
                     $log->U_Reproceso = 'N';
                     $log->save();
                 //borrar OP de control Piso
+               $orden_owor = NULL;
+               $orden_owor = DB::table('OWOR')->where('DocNum', Input::get('orden'))->first();
+            if ($orden_owor->PlannedQty == floatval ( $orden_owor->CmpltQty) ) {
                 DB::table('@CP_OF')->where('Code', Input::get('code'))->delete(); 
+            }
+                
                 //validar OP para cerrar 
                 $cerrar = DB::select("
                             SELECT T0.[DocNum] as Orden, T0.[ItemCode] as Codigo, T0.[PlannedQty] as Planeado, T0.[CmpltQty] as Terminado ,T0.UpdateDate as Actualizado FROM OWOR T0 LEFT JOIN (SELECT OWOR.DocNum as OP,sum(WOR1.PlannedQty) as Cantidad FROM OWOR inner join WOR1 on WOR1.DocEntry = OWOR.DocEntry inner join OITM A1 on WOR1.ItemCode=A1.ItemCode WHERE OWOR.[PlannedQty] <= OWOR.[CmpltQty] and  OWOR.[status] <> 'L' and WOR1.IssueType = 'M' and WOR1.IssuedQty < WOR1.PlannedQty and A1.ItmsGrpCod<> 113 group by OWOR.DocNum ) VAL on T0.DocNum = VAL.OP WHERE  T0.DocNum=? and T0.[PlannedQty] <=  T0.[CmpltQty] and  T0.[status] <> 'L' and VAL.Cantidad is null
@@ -1216,17 +1225,35 @@ public function terminarOP(){
                 if(count($cerrar) > 0){
                     SAP::ProductionOrderStatus(Input::get('orden'), 2); //cerrar Orden en SAP
                 }
-                Session::flash('mensaje', 'Recibo de Produccion Generado.');
+                Session::flash('mensaje', $result);
                 Session::put('recibo', 1);   
                 return redirect()->action('Mod01_ProduccionController@getOP', $id);
             } else {
                 //regresar el error de SAP 
                 Session::flash('error', $result);
                 return redirect()->action('Mod01_ProduccionController@getOP', $id);
-            } 
+            }
+            } else {
+                Session::flash('error', 'La cantidad Completada no puede ser mayor a la Planeada');
+                return redirect()->action('Mod01_ProduccionController@getOP', $id);
+            }
+           
+             
         }else{
             Session::flash('error', 'No estan capturados todos los "Tipos de Cambio" en SAP.');
                 return redirect()->action('Mod01_ProduccionController@getOP', $id);
+        }
+}
+public function getApellidoPaternoUsuario($apellido){
+        $preposiciones = ["DE", "LA", "LAS", "D", "LOS", "DEL"]; 
+        if (in_array($apellido[0], $preposiciones) && count($apellido)>1 ) {
+            if (in_array($apellido[1], $preposiciones) && count($apellido)>2 ) {
+               return $apellido[0].' '.$apellido[1].' '.$apellido[2];
+            } else {
+                return $apellido[0].' '.$apellido[1];
+            }            
+        } else {
+            return $apellido[0];
         }
 }
 }
